Fix CI: spellcheck, migration lint, Swift tests, duplicate WhoopBle method
- Add medical/scientific terms to cspell dictionary
- Use BIGINT for push_count in migration (squawk lint)
- Update readTypes count in Swift test (34 -> 57)
- Remove duplicate ensureCentralManager() from merge resolution

+++ drizzle/0049_clinical_records_and_new_metrics.sql
@@ -117,7 +117,7 @@ ALTER TABLE fitness.metric_stream
   ADD COLUMN IF NOT EXISTS electrodermal_activity REAL;
 
 ALTER TABLE fitness.daily_metrics
-  ADD COLUMN IF NOT EXISTS push_count INTEGER,
+  ADD COLUMN IF NOT EXISTS push_count BIGINT,
   ADD COLUMN IF NOT EXISTS wheelchair_distance_km REAL,
   ADD COLUMN IF NOT EXISTS uv_exposure REAL;
